get connection file from DUNEDAQ share path
