+++ scripts/hfButler.py
@@ -22,21 +22,6 @@ class HFObject(object):
 
 CONTEXT_SETTINGS = dict(help_option_names=['-h', '--help'])
 
-def find_conn_file():
-    path = os.environ.get('DUNEDAQ_SHARE_PATH')
-    name = 'dtp_connections.xml'
-    result = []
-
-    for dir in path.split(':'):
-        for root, ds, fs in os.walk(dir):
-            if name in fs:
-                result.append(os.path.join(root, name))
-
-    if len(result)>1:
-        print("Multiple connection files!")
-
-    return result[0]
-
 def get_devices(ctx, args, incomplete):
     devs = setup.connectionManager(ctx.params['connection']).getDevices()
     return [k for k in devs if incomplete in k]
@@ -45,7 +30,7 @@ extra_autocompl = {'autocompletion': get_devices} if parse_version(click.__versi
 
 @click.group(context_settings=CONTEXT_SETTINGS, invoke_without_command=True)
 @click.option('-e', '--exception-stack', 'aExcStack', is_flag=True, help="Display full exception stack")
-@click.option('-c', '--connection', type=click.Path(exists=True), default=find_conn_file())
+@click.option('-c', '--connection', type=click.Path(exists=True), default=setup.find_conn_file())
 @click.argument('device', **extra_autocompl)
 @click.pass_context
 @click.version_option(version='ultimate')
