+++ scripts/hfDriver.py
@@ -30,7 +30,7 @@ extra_autocompl = {'autocompletion': get_devices} if parse_version(click.__versi
 
 
 @click.command(context_settings=CONTEXT_SETTINGS)
-@click.option('-c', '--connection', type=click.Path(exists=True), envvar='CONNECTION_FILE')
+@click.option('-c', '--connection', type=click.Path(exists=True), default=setup.find_conn_file())
 @click.argument('device', **extra_autocompl)
 @click.option('--init/--no-init', 'init', help='Setup', default=False)
 @click.option('--dr-on/--dr-off', 'dr_on', help='Enable data-reception block', default=None)
