+++ scripts/hfWatcher.py
@@ -20,7 +20,7 @@ extra_autocompl = {'autocompletion': get_devices} if parse_version(click.__versi
 
 # -----------------------------------------------------------------------------
 @click.command(context_settings=CONTEXT_SETTINGS)
-@click.option('-c', '--connection', type=click.Path(exists=True), envvar='CONNECTION_FILE')
+@click.option('-c', '--connection', type=click.Path(exists=True), default=setup.find_conn_file())
 @click.argument('device', **extra_autocompl)
 @click.option('-r/-R', '--show-dr/--hide-dr', 'dr', default=True)
 @click.option('-b/-B', '--show-dpr/--hide-dpr', 'dpr', default=True)
