+++ scripts/hfWibulator.py
@@ -21,7 +21,7 @@ def get_devices(ctx, args, incomplete):
 extra_autocompl = {'autocompletion': get_devices} if parse_version(click.__version__) >= parse_version('7.0') else {}
 
 @click.command(context_settings=CONTEXT_SETTINGS)
-@click.option('-c', '--connection', type=click.Path(exists=True), envvar='CONNECTION_FILE')
+@click.option('-c', '--connection', type=click.Path(exists=True), default=setup.find_conn_file())
 @click.argument('device', **extra_autocompl)
 # Flowswitch
 @click.option('--src-sel', type=click.Choice(['gbt', 'wibtor']), help='Input source selection', default=None)
